changed some scripts
- deleted arrowButton scripts and included them in the selection manager scripts
