+++ Assets/Pixel_imgs/Tools/Chilled Martini Glass.png.meta
@@ -147,7 +147,8 @@ TextureImporter:
     secondaryTextures: []
     spriteCustomMetadata:
       entries: []
-    nameFileIdTable: {}
+    nameFileIdTable:
+      Chilled Martini Glass_0: 5786465989973217381
   mipmapLimitGroupName: 
   pSDRemoveMatte: 0
   userData:
