+++ Assets/Pixel_imgs/Tools/Chilled Stemmed Glass.png.meta
@@ -147,7 +147,8 @@ TextureImporter:
     secondaryTextures: []
     spriteCustomMetadata:
       entries: []
-    nameFileIdTable: {}
+    nameFileIdTable:
+      Chilled Stemmed Glass_0: 8644749224448089001
   mipmapLimitGroupName: 
   pSDRemoveMatte: 0
   userData:
