+++ Assets/Scenes/cardsScreen.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4a1707571331e4c1bad555e610fae1a4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
