+++ Assets/Scenes/cardsScreen/flashcards.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a68ee791fd252493c9508a3772ee2820
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
