--- Assets/Scenes/learnSelection/arrowButton.cs
@@ -1,13 +1,0 @@
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-namespace learnSelect
-{
-    public class arrowButton : MonoBehaviour
-    {
-        public void PreviousScreen()
-        {
-            SceneManager.LoadScene("Scenes/HomeScreen/homeScreen");
-        }
-    } 
-}
