--- Assets/Scenes/learnSelection/arrowButton.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 833c2c13857d84e548f241c0ae74225f
