+++ Assets/Scenes/learnSelection/learnSelect.unity
@@ -1878,7 +1878,6 @@ GameObject:
   - component: {fileID: 874666070}
   - component: {fileID: 874666069}
   - component: {fileID: 874666068}
-  - component: {fileID: 874666071}
   m_Layer: 5
   m_Name: arrowButton
   m_TagString: Untagged
@@ -1949,8 +1948,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 874666071}
-        m_TargetAssemblyTypeName: learnSelect.arrowButton, Assembly-CSharp
+      - m_Target: {fileID: 120179044}
+        m_TargetAssemblyTypeName: SelectionManager, Assembly-CSharp
         m_MethodName: PreviousScreen
         m_Mode: 1
         m_Arguments:
@@ -1999,18 +1998,6 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 874666065}
   m_CullTransparentMesh: 1
---- !u!114 &874666071
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 874666065}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 833c2c13857d84e548f241c0ae74225f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &943718444
 GameObject:
   m_ObjectHideFlags: 0
@@ -4508,7 +4495,7 @@ MonoBehaviour:
     m_SelectOnDown: {fileID: 0}
     m_SelectOnLeft: {fileID: 0}
     m_SelectOnRight: {fileID: 0}
-  m_Transition: 3
+  m_Transition: 1
   m_Colors:
     m_NormalColor: {r: 1, g: 1, b: 1, a: 1}
     m_HighlightedColor: {r: 0.9607843, g: 0.9607843, b: 0.9607843, a: 1}
