--- Assets/Scenes/profileSelection/arrowButton.cs
@@ -1,14 +1,0 @@
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-namespace ProfileSelection
-{
-    public class arrowButton : MonoBehaviour
-    {
-        public void PreviousScreen()
-        {
-            // Use scene name or scene index
-            SceneManager.LoadScene("Scenes/ProfilePage/profile");
-        }
-    }
-}
