--- Assets/Scenes/profileSelection/arrowButton.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 58d1f78b41b1648b994a99a6d47a1759
