+++ Assets/Scenes/profileSelection/buttonManager.cs
@@ -1,6 +1,7 @@
 using System;
 using UnityEngine;
 using UnityEngine.UI;
+using UnityEngine.SceneManagement;
 
 namespace profilePage
 {
@@ -11,6 +12,18 @@ namespace profilePage
         public string[] profiles = { "picture01", "picture02", "picture03", "picture04", "picture05", "picture06",
             "picture07", "picture08"};
 
+        public void selectedProfile()
+        {
+            
+            profile = grab the users document 
+                look for the profile field and return that number 
+                
+                number = the returned 
+                
+        }
+        
+        
+        
     private void Start()
         {
             // for testing
@@ -51,6 +64,12 @@ namespace profilePage
                     profilePictures[i].color = Color.white;
                 }
             }
+        }
+        
+        public void PreviousScreen()
+        {
+            // Use scene name or scene index
+            SceneManager.LoadScene("Scenes/ProfilePage/profile");
         }
     }
 }
