--- Assets/Scenes/theorySelection/arrowButton.cs
@@ -1,10 +1,0 @@
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class BackButton : MonoBehaviour
-{
-    public void Back()
-    {
-        SceneManager.LoadScene("Scenes/learnSelection/learnSelect");
-    }
-}
