--- Assets/Scenes/theorySelection/arrowButton.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 28dc877edf51a4192b4cdd393271e0a1
