+++ Assets/Scenes/theorySelection/theorySelect.unity
@@ -4611,7 +4611,6 @@ GameObject:
   - component: {fileID: 2049186995}
   - component: {fileID: 2049186994}
   - component: {fileID: 2049186993}
-  - component: {fileID: 2049186996}
   m_Layer: 5
   m_Name: arrowButton
   m_TagString: Untagged
@@ -4682,8 +4681,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 2049186996}
-        m_TargetAssemblyTypeName: BackButton, Assembly-CSharp
+      - m_Target: {fileID: 31126376}
+        m_TargetAssemblyTypeName: SelectManager, Assembly-CSharp
         m_MethodName: Back
         m_Mode: 1
         m_Arguments:
@@ -4732,18 +4731,6 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 2049186990}
   m_CullTransparentMesh: 1
---- !u!114 &2049186996
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2049186990}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 28dc877edf51a4192b4cdd393271e0a1, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &2054081497
 GameObject:
   m_ObjectHideFlags: 0
