+++ ProjectSettings/EditorBuildSettings.asset
@@ -44,5 +44,6 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/gameSelection/gameSelect.unity
     guid: bfcaf6d920d1448fcbd85283fd579dbe
-  m_configObjects: {}
+  m_configObjects:
+    com.unity.adaptiveperformance.simulator.provider_settings: {fileID: 11400000, guid: 9e10fd78c5f524fa79270245230bbfaf, type: 2}
   m_UseUCBPForAssetBundles: 0
